prepare final experiments

+++ main.py
@@ -64,11 +64,11 @@ class Args():
             self.parameter_shrink = 2
         else:
             self.parameter_shrink = 1
-        self.hidden_size_rnn = 128/self.parameter_shrink # hidden size for main RNN
+        self.hidden_size_rnn = int(128/self.parameter_shrink) # hidden size for main RNN
         self.hidden_size_rnn_output = 16 # hidden size for output RNN
-        self.embedding_size_rnn = 64/self.parameter_shrink # the size for LSTM input
+        self.embedding_size_rnn = int(64/self.parameter_shrink) # the size for LSTM input
         self.embedding_size_rnn_output = 4 # the embedding size for output rnn
-        self.embedding_size_output = 64/self.parameter_shrink # the embedding size for output (VAE/MLP)
+        self.embedding_size_output = int(64/self.parameter_shrink) # the embedding size for output (VAE/MLP)
 
         self.batch_size = 32 # normal: 64, and the rest should be changed accordingly
         self.test_batch_size = 32
